LayoutTests/http/tests/security/xss-DENIED-frame-name-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL about:blank from frame with URL http://localhost:8000/security/resources/frame-for-parent-name.html. Domains, protocols and ports must match.

ALERT: undefined
   This test passes if it doesn't alert fail.
